refactor(runtime): 将 RuntimeModalContent 内联合并至 RuntimeSettings 以消除中间组件层
